fix: money 정수값 범위를 벗어나면 에러 해결
long으로 변경

+++ src/main/java/lotto/domain/Buyer.java
@@ -24,11 +24,11 @@ public class Buyer {
         this.money = money;
     }
 
-    public int getMoney() {
+    public long getMoney() {
         return this.money.getMoney();
     }
 
-    public int getLottoCount() {
+    public long getLottoCount() {
         return this.money.getLottoCount();
     }
 }
